requirements.txt
fastapi
uvicorn
sqlalchemy
pyperclip
pydantic
python-dotenv
